requirements.txt
fastapi
pydantic==1.9.0
uvicorn==0.18.3
python-decouple==3.6
requests==2.28.1